matz
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@1053 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ruby.h
@@ -380,9 +380,9 @@ struct RBignum {
 #define xrealloc ruby_xrealloc
 #define xfree ruby_xfree
 
-void *xmalloc _((size_t));
-void *xcalloc _((size_t,size_t));
-void *xrealloc _((void*,size_t));
+void *xmalloc _((long));
+void *xcalloc _((long,long));
+void *xrealloc _((void*,long));
 void xfree _((void*));
 
 #define ALLOC_N(type,n) (type*)xmalloc(sizeof(type)*(n))
